+++ src/main/java/it/eng/idsa/businesslogic/service/DapsTokenProviderService.java
@@ -12,4 +12,6 @@ public interface DapsTokenProviderService {
 	 * @return
 	 */
 	DynamicAttributeToken getDynamicAtributeToken();
+
+	String getConnectorUUID();
 }