+++ src/main/java/it/eng/idsa/businesslogic/service/impl/CommunicationServiceImpl.java
@@ -56,5 +56,26 @@ public class CommunicationServiceImpl implements CommunicationService {
 		
 		return result.getBody();
 	}
+	
+	@Override
+	public void deleteRequest(String endpoint) {
+		try {
+			restTemplate.delete(endpoint);
+		} catch (Exception e) {
+			logger.error("Error while making a delete request", e);
+		}
+	}
+
+	@Override
+	public String getRequest(String endpoint) {
+		ResponseEntity<String> result;
+		try {
+			result = restTemplate.getForEntity(endpoint, String.class);
+		} catch (Exception e) {
+			logger.error("Error while making a request", e);
+			return null;
+		}
+		return result.getBody();
+	}
 
 }